Implemented Works

+++ config/initializers/sanitize.rb
@@ -78,7 +78,7 @@ class Sanitize
       node_name = env[:node_name]
 
       if node_name == "iframe"
-        unless node['src'].include? "youtube.com"
+        unless node['src'].include? "youtube.com" && "nicovideo.jp"
           node.unlink
           {:whitelist_nodes => [node]}
         end
